requirements.txt
Beaker==1.6.4
Django==1.5.5
Django-Select2==4.2.2
Flask==0.10.1
Jinja2==2.7.1
Markdown==2.3.1
MarkupSafe==0.18
MySQL-python==1.2.4
PIL==1.1.7
PyYAML==3.10
South==0.8.2
Unidecode==0.04.14
Werkzeug==0.9.4
coverage==3.6
distribute==0.6.14
django-admin-bootstrapped==0.4.3
django-ajax-selects==1.2.5
django-autocomplete-light==1.4.3
django-autofixture==0.3.1
django-autoslug==1.7.1
django-axes==1.3.6
django-bootstrap-pagination==0.1.11
django-bootstrap-toolkit==2.15.0
django-bootstrap3-datetimepicker==1.0.3
django-bower==4.7
django-cities==0.2
django-cities-light==2.1.1
django-contact-form==1.0
django-dajax==0.9.2
django-dajaxice==0.5.5
django-datatable-view==0.6
django-datatables-view==1.6
django-debug-toolbar==0.9.4
django-eztables==0.3.2
django-filter==0.7
django-floppyforms==1.1
django-jinja==0.21
django-json-field==0.5.5
django-like==0.0.7
django-messages==0.4.4
django-mockups==0.4.8
django-model-utils==1.5.0
django-multi-model-form==1.1.0
django-password-reset==0.5.1
django-postman==3.0.2
django-profiles==0.2
django-registration==1.0
django-session-security==2.1.2
django-smart-selects==1.0.9
django-suit==0.2.5
django-transmeta==0.7.1
django.js==0.8.1
django-rest-swagger
django-rest-framework-docs
django-postal
django-uuidfield
django-push-notifications
djangorestframework==2.3.8
docutils==0.11
identicon==0.0.0
itsdangerous==0.23
mock==1.0.1
progressbar==2.3dev
pyparsing==2.0.1
python-dateutil==2.1
querystring-parser==1.1
six==1.4.1
suit==2.0.2
textile==2.1.5
wsgiref==0.1.2
